Create `saxaboom-runtime` crate with bindings to runtime APIs (#17)
We've long ran around with hand-written `repr(C)` structures and
constants in our own codebase, also causing a lot of headaches when some
layouts and values changed in the `1.0` -> `1.1` header upgrade without
being visible.

Instead, by generating a crate with types from the actual headers, we
will immediately see differences and incompatibilities when new headers
are incorporated in the form of compiler errors, and as a bonus expose
all our efforts in wrapping the structures and **some** helper functions
to everyone else using `saxaboom`.

Note that some effort has been taken to exclude bindings to the helper
functions defined in the library as they pass `id<>` structures around
which is not something a simple C API via `bindgen` can represent. You
can find a head-start for this at
https://github.com/Traverse-Research/saxaboom/compare/runtime-functions
but it won't have a sensible ABI, nor does bindgen seem to generate
getters for symbols.