runtime: Disable default `link` feature on `metal` crate
Since we're only using a few types from the `metal` crate, this helper
library has no reason or need to link against framework libraries (i.e.
we don't call global static functions on them).  Disable the `link`
feature on the `metal` crate, but note that for now this won't disable
it on the `core-graphics-types` crate until
https://github.com/gfx-rs/metal-rs/pull/330 is merged.