Upgrade to `objc2 0.6` and `objc2-metal 0.3` (#31)
`objc2-metal` has a new `to_raw()` function for converting
`MTLResourceID` to its raw representation, as used in a few structures
as a raw integer.

Note also that a new `MTLAllocation` interface was added to Metal. Even
if that interface is _optional_, current `objc2` design requires this
trait to be part of the hierarchy (and the feature explicitly enabled)
to get access to all "descendant" interfaces (`MTLResource`, which
provides `MTLBuffer` and `MTLTexture`). There are upstream plans to see
if this can be simplified: https://github.com/madsmtm/objc2/issues/663.