Add reflection info for all shader stages (#38)
Similar to
[`compute_info`](https://github.com/Traverse-Research/saxaboom/blob/ee58c32ec432b694168266b1e1510f6e806089bc/src/lib.rs#L125),
this PR adds reflection info for other shader stages. This is useful for
mesh/amplification shaders, because Metal needs to know the number of
thread groups to spawn as well as `needs_draw_params`, which can be used
in conjunction with
https://github.com/Traverse-Research/saxaboom/pull/37

And because there are a lot of shader stages, it's quite tedious to
write same variation of "Info" structure for every single one, so I
added a macro.

---------

Co-authored-by: Marijn Suijten <marijn@traverseresearch.nl>