runtime: Fix and require compatibility with `objc2-metal 0.3.2` (#44)
Technically a breaking change, `.gpuAddress()` is now behind a new
`MTLGPUAddress` crate feature that didn't exist on `0.3.1` yet,
requiring us to add the flag and bump the minimum version bound to
`0.3.2`.

Alternatively downstream crates depend on `objc2-metal 0.3.2` with the
`MTLGPUAddress` feature that enables it to be used here, which is why we
didn't notice it yet.

CC @madsmtm